requirements-dev.txt
# ============================================================================
# PHENO-SDK INTEGRATIONS (Local Development)
# ============================================================================
# Install local mcp-QA package for structured logging, OAuth, and test framework
-e ../pheno-sdk/mcp-QA
-e ../pheno-sdk/KInfra/libraries/python
-e ../pheno-sdk/deploy-kit
-e ../pheno-sdk/observability-kit
-e ../pheno-sdk/cli-builder-kit
-e ../pheno-sdk/filewatch-kit
-e ../pheno-sdk/pydevkit
-e ../pheno-sdk/db-kit
-e ../pheno-sdk/authkit-client
-e ../pheno-sdk/adapter-kit
-e ../pheno-sdk/vector-kit
-e ../pheno-sdk/workflow-kit

# ============================================================================
# CORE DEPENDENCIES
# ============================================================================
fastmcp>=2.12.2
supabase>=2.5.0
python-dotenv>=1.0.1
google-cloud-aiplatform>=1.49.0
aiohttp>=3.8.0
workos>=1.0.0
PyJWT>=2.8.0
cryptography>=41.0.0
tqdm>=4.66.0
rapidfuzz>=3.10.0

# ============================================================================
# CORE UTILITIES
# ============================================================================
pydantic[email]>=2.11.7,<3.0.0
httpx>=0.28.1,<1.0.0
PyYAML>=6.0
psutil>=5.9.0
typing-extensions>=4.12.2

# ============================================================================
# TESTING & QA
# ============================================================================
pytest>=9.0.0,<10.0.0  # Pin to 9.x for compatibility
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0
pytest-xdist>=3.6.1
pytest-dependency>=0.6.0
coverage>=7.4.0
playwright>=1.40.0  # For automated OAuth flow (get_authkit_token_playwright.py)
pytest-playwright>=0.4.0  # Pytest integration for Playwright (better test integration)
keyring>=25.0.0  # OS keychain access for secure token storage

# ============================================================================
# CODE QUALITY
# ============================================================================
black>=23.0.0
ruff>=0.1.0
isort>=5.12.0
